Initial Restructuring (#47)
Changed the structure of "Home","Overview" and "Tutorial Introduction" to fit the content and minimize scrolling.
Created a two-column design for those pages to do so.

Sourced the glossary (previously embedded into the Tutorial Introduction text) out into it's own entry in the navbar and created a collapsable, sticky element that stays next to the Tutorial Introduction text to improve usability. Split the glossary into three tables with separate headings and put all contents into a partial for both glossary pages to be maintained easily. Also added links to keywords in the glossary.

Edited a lot of the pages content in regards of grammar, removed redundant information and summarized. 

Also used alert classes for the content to have a clearer structure.